% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/06_get_absence.R
\name{get_absence}
\alias{get_absence}
\title{get_absence()}
\usage{
get_absence(student_rcard)
}
\arguments{
\item{student_rcard}{A vector containing a student's report card information}
}
\description{
This function takes a student's report card as input (vector), and extract the number of absences
}
\examples{
parsed_pdf <-read_pdf('my_pdf_file.pdf')
students_list <- split_students(parsed_pdf)
get_absence(students_list[[1]])
}
\keyword{student_rcard}

