% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{get_cumgpa}
\alias{get_cumgpa}
\title{get_cumgpa()}
\usage{
get_cumgpa(student_rcard)
}
\arguments{
\item{student_rcard}{A vector containing a student's report card information}
}
\description{
This function takes a student's report card as input (vector), and extract the students cumulative gpa
}
\examples{
parsed_pdf <-read_pdf('my_pdf_file.pdf')
students_list <- split_students(parsed_pdf)
get_cumgpa(students_list[[1]])
}
\keyword{student_rcard}

