% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/09_get_date.R
\name{get_date}
\alias{get_date}
\title{get_date()}
\usage{
get_date(student_rcard)
}
\arguments{
\item{student_rcard}{A vector containing a student's report card information}
}
\description{
This function takes a student's report card as input (vector), and extract the term end's date
}
\examples{
parsed_pdf <-read_pdf('my_pdf_file.pdf')
students_list <- split_students(parsed_pdf)
get_date(students_list[[1]])
}
\keyword{student_rcard}

