% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/04_get_gpa.R
\name{get_gpa}
\alias{get_gpa}
\title{get_gpa()}
\usage{
get_gpa(student_rcard)
}
\arguments{
\item{student_rcard}{A vector containing a student's report card information}
}
\description{
This function takes a student's report card as input (vector), and extract the students gpas
}
\examples{
parsed_pdf <-read_pdf('my_pdf_file.pdf')
students_list <- split_students(parsed_pdf)
get_gpa(students_list[[1]])
}
\keyword{student_rcard}

