% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{get_grade}
\alias{get_grade}
\title{get_grade()}
\usage{
get_grade(student_rcard)
}
\arguments{
\item{student_rcard}{A vector containing a student's report card information}
}
\description{
This function takes a student's report card as input (vector), and extract the student's grade
}
\examples{
parsed_pdf <-read_pdf('my_pdf_file.pdf')
students_list <- split_students(parsed_pdf)
get_grade(students_list[[1]])
}
\keyword{student_rcard}

