% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/11_get_grades.R
\name{get_grades}
\alias{get_grades}
\title{get_grades()}
\usage{
get_grades(student_rcard)
}
\arguments{
\item{student_rcard}{A vector containing a student's report card information}
}
\description{
This function takes a student's report card as input (vector), and extract the student's grades and courses.
}
\examples{
parsed_pdf <-read_pdf('my_pdf_file.pdf')
students_list <- split_students(parsed_pdf)
student_rcard <- students_list[[1]]
get_grades(student_rcard)
}
\keyword{student_rcard}

