% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/08_get_id.R
\name{get_id}
\alias{get_id}
\title{get_id()}
\usage{
get_id(student_rcard)
}
\arguments{
\item{student_rcard}{A vector containing a student's report card information}
}
\description{
This function takes a student's report card as input (vector), and extract the student's id
}
\examples{
parsed_pdf <-read_pdf('my_pdf_file.pdf')
students_list <- split_students(parsed_pdf)
get_id(students_list[[1]])
}
\keyword{student_rcard}

