% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/03_get_name.R
\name{get_name}
\alias{get_name}
\title{get_name()}
\usage{
get_name(student_rcard)
}
\arguments{
\item{student_rcard}{A vector containing a student's report card information}
}
\description{
This function takes a student's report card as input (vector), and extract the students first and last name.
}
\examples{
parsed_pdf <-read_pdf('my_pdf_file.pdf')
students_list <- split_students(parsed_pdf)
get_name(students_list[[1]])
[[1]]
[1] "First name" "Last name"
}
\keyword{student_rcard}

