% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{read_pdf}
\alias{read_pdf}
\title{read_pdf()}
\usage{
read_pdf(file, xpdf_options = "-table -eol dos")
}
\arguments{
\item{file}{The name of the file which the data are to be read from (including the .pdf extension). read_pdf that the file to be read from is in the working directory}

\item{xpdf_options}{Specify options for pdftotext software. Refer to http://linux.die.net/man/1/pdftotext for the full list of available options.}
}
\description{
This function takes a DC report card (pdf) as input, reads its content and output this content as text
}
\examples{
read_pdf('my_pdf_file.pdf')
}
\keyword{my_function}

